refactor: extract helpers and reformulation mixin from tool_retriever.py
